Refactor CodeEditor component and its components
