+++ src/components/CodeEditor/index.jsx
@@ -1,9 +1,7 @@
-import React, { useRef, useEffect, useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import T from 'prop-types'
 import cx from 'classnames'
 import { connect } from 'react-redux'
-import CodeMirror from 'codemirror'
-import jsonlint from 'jsonlint-mod'
 import beautify from 'js-beautify'
 import debounce from 'lodash.debounce'
 
@@ -20,114 +18,34 @@ import {
   selectResponseWidth,
 } from 'flux/modules/user'
 import { MIN_TEXTAREA_WIDTH } from 'dictionary'
-import { checkIsResponseError } from 'helpers'
+import {
+  checkIsResponseError,
+  initCodeEditor,
+  getEditorSize,
+  getEditorErrors,
+} from 'helpers'
 import { ReactComponent as DragIconComponent } from 'assets/drag.svg'
 import Textarea from './Textarea'
 import Actions from './Actions'
 
-import 'codemirror/addon/selection/mark-selection'
-import 'codemirror/addon/edit/closebrackets'
-import 'codemirror/mode/javascript/javascript'
-import 'codemirror/addon/lint/lint'
-import 'codemirror/addon/lint/json-lint'
-import 'codemirror/lib/codemirror.css'
-import 'codemirror/addon/lint/lint.css'
 import './index.scss'
 
-window.jsonlint = jsonlint
-
-function initCodeEditor(
-  requestDomEl,
-  responseDomEl,
-  requestWidth,
-  responseWidth,
-  setRequestEditor,
-  setResponseEditor
-) {
-  const requestEditor = CodeMirror.fromTextArea(
-    requestDomEl,
-    {
-      mode: { name: 'javascript', json: true },
-      theme: 'textarea',
-      lineWrapping: true,
-      autoCloseBrackets: true,
-      lint: {
-        lintOnChange: true,
-        selfContain: false,
-      },
-      selfContain: false,
-    }
-  )
-  requestEditor.setSize('100%', null)
-  // eslint-disable-next-line no-param-reassign
-  requestDomEl.parentElement.style.width = requestWidth
-    ? `${requestWidth}px`
-    : '50%'
-
-  const responseEditor = CodeMirror.fromTextArea(
-    responseDomEl,
-    {
-      mode: {
-        name: 'javascript',
-        json: true,
-      },
-      theme: 'textarea',
-      readOnly: 'nocursor',
-      lineWrapping: true,
-      styleSelectedText: true,
-    }
-  )
-  responseEditor.setSize('100%', null)
-  // eslint-disable-next-line no-param-reassign
-  responseDomEl.parentElement.style.width = responseWidth
-    ? `${responseWidth}px`
-    : '50%'
-
-  setRequestEditor(requestEditor)
-  setResponseEditor(responseEditor)
-}
-
-function getEditorSize(editorDomEl) {
-  const wrapDomEl = editorDomEl.parentElement
-  const editorDomRect = wrapDomEl.getBoundingClientRect()
-  const { width } = editorDomRect
-  const { height } = editorDomRect
-
-  return {
-    width,
-    height,
-    wrapDomEl,
-  }
-}
-
-function getEditorErrors(editorInstance) {
-  editorInstance.performLint()
-
-  const value = editorInstance.getValue()
-  return CodeMirror.lint.json(value) || []
-}
-
-function CodeEditor(props) {
-  const dragEl = useRef(null)
+function CodeEditor({
+  requestText,
+  responseText,
+  makeRequest,
+  className,
+  isLoading,
+  savedRequestWidth,
+  savedResponseWidth,
+  saveRequestWidth,
+  saveResponseWidth,
+}) {
   const [requestEditor, setRequestEditor] = useState(null)
   const [responseEditor, setResponseEditor] = useState(null)
   const [isLintError, setIsLintError] = useState(false)
 
-  const {
-    requestText,
-    responseText,
-    makeRequest,
-    className,
-    isLoading,
-    savedRequestWidth,
-    savedResponseWidth,
-    saveRequestWidth,
-    saveResponseWidth,
-  } = props
-
-  const isResponseError = checkIsResponseError(responseText)
-
-  // INIT
+  // INIT FROM LOCAL STORAGE
   useEffect(() => {
     if (requestEditor || responseEditor) {
       return
@@ -199,7 +117,7 @@ function CodeEditor(props) {
 
   // RESIZE ON MOUSE MOVE
   useEffect(() => {
-    const dragDomEl = dragEl.current
+    const dragDomEl = document.getElementById('drag-el')
     const requestTextareaDomEl = document.getElementById(
       'request-textarea'
     )
@@ -292,21 +210,23 @@ function CodeEditor(props) {
       const nextResponseWidth =
         editorsWidth - nextRequestWidth
       if (
-        nextRequestWidth > MIN_TEXTAREA_WIDTH &&
-        nextResponseWidth > MIN_TEXTAREA_WIDTH
+        nextRequestWidth < MIN_TEXTAREA_WIDTH ||
+        nextResponseWidth < MIN_TEXTAREA_WIDTH
       ) {
-        requestWrapDomEl.style.width = `${nextRequestWidth}px`
-        responseWrapDomEl.style.width = `${nextResponseWidth}px`
-        setRequestWidth(nextResponseWidth)
-        setResponseWidth(nextResponseWidth)
+        return
       }
+
+      requestWrapDomEl.style.width = `${nextRequestWidth}px`
+      responseWrapDomEl.style.width = `${nextResponseWidth}px`
+      saveRequestWidth(nextRequestWidth)
+      saveResponseWidth(nextResponseWidth)
     }
 
     window.addEventListener('resize', onResize)
     return () => {
       window.removeEventListener('resize', onResize)
     }
-  }, [])
+  }, [saveRequestWidth, saveResponseWidth])
 
   const onBeautify = () => {
     requestEditor.setValue(
@@ -320,17 +240,13 @@ function CodeEditor(props) {
     }
   }
 
-  const classNames = cx({
-    'code-editor': true,
-    [className]: className,
-  })
-  const editorTextareaCl = cx({
-    'code-editor__textarea': true,
-    'code-editor__editor_error': isResponseError,
-  })
-  const dragIconCl = cx({
-    'code-editor__drag-icon': true,
-  })
+  const isResponseError = checkIsResponseError(responseText)
+  const classNames = cx(['code-editor', className])
+  const editorTextareaCl = cx([
+    'code-editor__textarea',
+    isResponseError && 'code-editor__editor_error',
+  ])
+  const dragIconCl = cx(['code-editor__drag-icon'])
 
   return (
     <div className={classNames}>
@@ -342,7 +258,7 @@ function CodeEditor(props) {
           id="request-textarea"
           isError={isResponseError}
         />
-        <div ref={dragEl} className="code-editor__drag">
+        <div id="drag-el" className="code-editor__drag">
           <DragIconComponent className={dragIconCl} />
         </div>
         <Textarea
