+++ src/helpers/getEditorErrors.js
@@ -1,0 +1,10 @@
+import CodeMirror from 'codemirror'
+
+function getEditorErrors(editorInstance) {
+  editorInstance.performLint()
+
+  const value = editorInstance.getValue()
+  return CodeMirror.lint.json(value) || []
+}
+
+export default getEditorErrors
