+++ src/helpers/getEditorSize.js
@@ -1,0 +1,14 @@
+function getEditorSize(editorDomEl) {
+  const wrapDomEl = editorDomEl.parentElement
+  const editorDomRect = wrapDomEl.getBoundingClientRect()
+  const { width } = editorDomRect
+  const { height } = editorDomRect
+
+  return {
+    width,
+    height,
+    wrapDomEl,
+  }
+}
+
+export default getEditorSize
