+++ src/helpers/index.js
@@ -3,6 +3,9 @@ import getResponseError from './getResponseError'
 import checkIsResponseError from './checkIsResponseError'
 import pluralizeWord from './pluralizeWord'
 import deepClone from './deepClone'
+import initCodeEditor from './initCodeEditor'
+import getEditorSize from './getEditorSize'
+import getEditorErrors from './getEditorErrors'
 import getRandomId from './getRandomId'
 import RequestsHistory from './RequestsHistory'
 import getRequestId from './getRequestId'
@@ -13,6 +16,9 @@ import isNumeric from './isNumeric'
 export {
   isNumeric,
   getRandomId,
+  initCodeEditor,
+  getEditorSize,
+  getEditorErrors,
   checkIsResponseError,
   getResponseError,
   pluralizeWord,
